refactor: @LoginMember와 resolver를 이용한 컨트롤러 리팩토링(중복 코드 제거)
