+++ src/main/java/com/saebyeok/saebyeok/controller/AnalysisController.java
@@ -1,15 +1,12 @@
 package com.saebyeok.saebyeok.controller;
 
+import com.saebyeok.saebyeok.controller.resolver.LoginMember;
 import com.saebyeok.saebyeok.domain.Member;
-import com.saebyeok.saebyeok.domain.MemberRepository;
 import com.saebyeok.saebyeok.dto.ArticlesAnalysisResponse;
 import com.saebyeok.saebyeok.dto.CommentsAnalysisResponse;
-import com.saebyeok.saebyeok.exception.MemberNotFoundException;
-import com.saebyeok.saebyeok.security.User;
 import com.saebyeok.saebyeok.service.AnalysisService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,37 +17,26 @@ import java.util.List;
 @RequiredArgsConstructor
 @RestController
 public class AnalysisController {
-    private final MemberRepository memberRepository;
     private final AnalysisService analysisService;
 
     @GetMapping("/analysis/articles")
-    public ResponseEntity<ArticlesAnalysisResponse> getArticlesAnalysis(Authentication authentication) {
-        // TODO: 2020/08/16 커스텀 어노테이션으로 리팩토링
-        User user = (User) authentication.getPrincipal();
-        Member member = memberRepository.findById(user.getId())
-                .orElseThrow(() -> new MemberNotFoundException(user.getId()));
-
+    public ResponseEntity<ArticlesAnalysisResponse> getArticlesAnalysis(@LoginMember Member member) {
         List<Integer> articleEmotionsCount = analysisService.findArticleEmotionsCount(member);
         Long mostEmotionId = analysisService.findMostEmotionId(member);
 
         ArticlesAnalysisResponse articlesAnalysisResponse = new ArticlesAnalysisResponse(articleEmotionsCount,
-                                                                                         mostEmotionId);
+                mostEmotionId);
 
         return ResponseEntity.ok(articlesAnalysisResponse);
     }
 
     @GetMapping("/analysis/comments")
-    public ResponseEntity<CommentsAnalysisResponse> getCommentsAnalysis(Authentication authentication) {
-        // TODO: 2020/08/16 커스텀 어노테이션으로 리팩토링
-        User user = (User) authentication.getPrincipal();
-        Member member = memberRepository.findById(user.getId())
-                .orElseThrow(() -> new MemberNotFoundException(user.getId()));
-
+    public ResponseEntity<CommentsAnalysisResponse> getCommentsAnalysis(@LoginMember Member member) {
         Long totalCommentsCount = analysisService.countTotalCommentsBy(member);
         Long totalCommentLikesCount = analysisService.countTotalCommentLikesBy(member);
 
         CommentsAnalysisResponse commentsAnalysisResponse = new CommentsAnalysisResponse(totalCommentsCount,
-                                                                                         totalCommentLikesCount);
+                totalCommentLikesCount);
 
         return ResponseEntity.ok(commentsAnalysisResponse);
     }
